a workable data is added
The GIS data needed to successfully run the simulation are added. The three flood maps are identical and they have very coarse resolution (1km). The polygon shapefiles that generate the agents have only five features each. The purpose of adding these files is to successfully run the model with no error. The data will not replicate what is presented in Abebe et al 2020.